finish api to handle native side sql read/write

+++ android/app/src/main/kotlin/com/example/test_env/model/WeatherInfo.kt
@@ -2,17 +2,17 @@ package com.example.test_env.model
 
 import org.json.JSONObject
 
-data class WeatherInfo(var location: String = "undefined", var country: String = "undefined", var lastUpdate: String = "undefined", var tempC: Float = 0.0f, var windDegree: Int = 0, var windDir: String = "undefined", var cloud: Int = 0, var uv: Float = 0.0f) {
+data class WeatherInfo(var location: String = "undefined", var country: String = "undefined", var lastUpdate: String = "undefined", var tempC: Double = 0.0, var windDegree: Int = 0, var windDir: String = "undefined", var cloud: Int = 0, var uv: Double = 0.0) {
     companion object{
         fun jsonParse(obj: JSONObject):WeatherInfo {
             val location = obj["location"] as String
             val country = obj["country"] as String
             val lastUpdate = obj["last_update"] as String
-            val tempC = obj["temp_c"] as Float
+            val tempC = obj["temp_c"] as Double
             val windDegree = obj["wind_degree"] as Int
             val windDir = obj["wind_dir"] as String
             val cloud = obj["cloud"] as Int
-            val uv = obj["uv"] as Float
+            val uv = obj["uv"] as Double
 
             return WeatherInfo(location, country, lastUpdate, tempC, windDegree, windDir, cloud, uv)
         }
